requirements.txt

jaclang
jac-cloud
jac-streamlit
requests
byllm
js2py
google-generativeai
